[BE] Feat: tiny 인스턴스 사이즈 추가 - tiny 사이즈 추가 (0.25 CPU, 256MB RAM, 5GB 디스크) - CPU 소수점 지원 (0.25, 0.5 등) - DB 스키마 및 Go 모델 CPU 타입 변경 - K8s CRD 호환을 위한 CPU string 변환

+++ backend/internal/platform/store/postgres/migrations/010_db_instances.sql
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS db_instances
     created_from_preset   VARCHAR(100),
 
     -- 리소스 (프리셋에서 복사되거나 커스텀 입력)
-    cpu                   INTEGER CHECK (cpu > 0)            NOT NULL,
+    cpu                   DECIMAL(4, 2) CHECK (cpu > 0)      NOT NULL,
     memory                INTEGER CHECK (memory > 0)         NOT NULL, -- MB
     disk                  INTEGER CHECK (disk > 0)           NOT NULL, -- GB
